+++ .github/labeler.yml
@@ -21,8 +21,8 @@
 - 'src/cpp/src/llm_pipeline.cpp'
 - 'src/cpp/src/lm_encoding.hpp'
 - 'src/cpp/src/lm_encoding.cpp'
-- 'src/cpp/src/llm_pipeline_base.hpp'
-- 'src/cpp/src/llm_pipeline_static.hpp'
+- 'src/cpp/src/llm/pipeline_base.hpp'
+- 'src/cpp/src/llm/pipeline_static.hpp'
 - 'src/cpp/src/llm_pipeline_static.cpp'
 - 'src/cpp/src/text_streamer.cpp'
 - 'src/python/py_llm_pipeline.cpp'
@@ -44,9 +44,9 @@
 - 'src/cpp/include/openvino/genai/lora_adapter.hpp'
 - 'src/cpp/src/lora_adapter.cpp'
 - 'src/cpp/src/lora_helper.cpp'
-- 'src/cpp/src/lora_helper.hpp'
+- 'src/cpp/src/helper.hpp'
 - 'src/cpp/src/lora_names_mapping.cpp'
-- 'src/cpp/src/lora_names_mapping.hpp'
+- 'src/cpp/src/names_mapping.hpp'
 - 'src/cpp/src/safetensors.c'
 - 'src/python/py_lora_adapter.cpp'
 
@@ -58,10 +58,10 @@
 - 'src/cpp/include/openvino/genai/whisper_pipeline.hpp'
 - 'src/cpp/src/whisper/**/*'
 - 'src/cpp/src/whisper_generation_config.cpp'
-- 'src/cpp/src/whisper_pipeline_base.hpp'
+- 'src/cpp/src/whisper/pipeline_base.hpp'
 - 'src/cpp/src/whisper_pipeline.cpp'
 - 'src/cpp/src/whisper_pipeline_static.cpp'
-- 'src/cpp/src/whisper_pipeline_static.hpp'
+- 'src/cpp/src/whisper/pipeline_static.hpp'
 - 'src/python/py_whisper_pipeline.cpp'
 - 'tests/python_tests/test_whisper_pipeline.py'
 - 'tests/python_tests/test_whisper_pipeline_static.py'
