+++ src/cpp/src/continuous_batching/pipeline_base.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2023-2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "continuous_batching/icontinuous_batching.hpp"
+#include "continuous_batching/pipeline_base.hpp"
 
 namespace {
 std::string add_image_tags_to_prompt(const std::string& prompt, const std::vector<ov::Tensor>& rgbs, size_t history_images_size) {
