+++ src/cpp/src/image_generation/models/autoencoder_kl.cpp
@@ -17,7 +17,7 @@
 #include "utils.hpp"
 
 #include "json_utils.hpp"
-#include "lora/lora_helper.hpp"
+#include "lora/helper.hpp"
 
 namespace ov {
 namespace genai {
