+++ src/cpp/src/image_generation/models/clip_text_model.cpp
@@ -6,7 +6,7 @@
 #include <fstream>
 
 #include "json_utils.hpp"
-#include "lora/lora_helper.hpp"
+#include "lora/helper.hpp"
 #include "utils.hpp"
 
 namespace ov {
