+++ src/cpp/src/image_generation/models/flux_transformer_2d_model.cpp
@@ -7,7 +7,7 @@
 
 #include "json_utils.hpp"
 #include "utils.hpp"
-#include "lora/lora_helper.hpp"
+#include "lora/helper.hpp"
 
 namespace ov {
 namespace genai {
