+++ src/cpp/src/image_generation/models/unet2d_condition_model.cpp
@@ -8,7 +8,7 @@
 #include <fstream>
 
 #include "json_utils.hpp"
-#include "lora/lora_helper.hpp"
+#include "lora/helper.hpp"
 #include "utils.hpp"
 
 namespace ov {
