+++ src/cpp/src/image_generation/models/unet_inference_static_bs1.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "lora/lora_helper.hpp"
+#include "lora/helper.hpp"
 #include "image_generation/models/unet_inference.hpp"
 #include "utils.hpp"
