+++ src/cpp/src/llm/pipeline_base.hpp
