+++ src/cpp/src/llm/pipeline_static.cpp
@@ -1,9 +1,9 @@
 // Copyright (C) 2024-2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "llm_pipeline_static.hpp"
+#include "llm/pipeline_static.hpp"
 
-#include "sampler.hpp"
+#include "sampling/sampler.hpp"
 #include "utils.hpp"
 
 #include <fstream>
