+++ src/cpp/src/prompt_lookup/prompt_lookup_impl.hpp
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "openvino/genai/continuous_batching_pipeline.hpp"
-#include "continuous_batching/continuous_batching_impl.hpp"
+#include "continuous_batching/pipeline_impl.hpp"
 #include "continuous_batching_for_prompt_lookup.hpp"
 #include "speculative_decoding/speculative_decoding_metrics.hpp"
 #include "utils.hpp"
