+++ src/cpp/src/sampling/logit_processor.hpp
