+++ src/cpp/src/speculative_decoding/continuous_batching_for_speculative_decoding_impl.hpp
@@ -5,7 +5,7 @@
 
 #include "openvino/genai/continuous_batching_pipeline.hpp"
 
-#include "continuous_batching/continuous_batching_impl.hpp"
+#include "continuous_batching/pipeline_impl.hpp"
 #include "speculative_decoding/update_request_structs.hpp"
 
 namespace ov::genai {
