+++ src/cpp/src/whisper/config.hpp
