+++ src/cpp/src/whisper/feature_extractor.hpp
