+++ src/cpp/src/whisper/generation_config.cpp
