+++ src/cpp/src/whisper/models.hpp
