+++ tests/cpp/sampler.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <gtest/gtest.h>
-#include "sampler.hpp"
+#include "sampling/sampler.hpp"
 #include "openvino/genai/generation_config.hpp"
